Delete ViewCell class and move logic to AbstractViewCell
