+++ viewcelladapter/src/main/java/ca/antonious/viewcelladapter/viewcells/AbstractViewCell.java
@@ -2,6 +2,9 @@ package ca.antonious.viewcelladapter.viewcells;
 
 import android.view.View;
 
+import java.lang.reflect.Constructor;
+import java.lang.reflect.ParameterizedType;
+
 import ca.antonious.viewcelladapter.Function;
 
 /**
@@ -14,6 +17,10 @@ public abstract class AbstractViewCell<TViewHolder extends BaseViewHolder> imple
     public AbstractViewCell() {
         this.isSelected = false;
     }
+
+    public abstract int getLayoutId();
+    public abstract int getItemId();
+    public abstract void bindViewCell(TViewHolder viewHolder);
 
     @Override
     public boolean isSelected() {
@@ -30,8 +37,31 @@ public abstract class AbstractViewCell<TViewHolder extends BaseViewHolder> imple
         isSelected = false;
     }
 
-    public abstract int getLayoutId();
-    public abstract int getItemId();
-    public abstract void bindViewCell(TViewHolder viewHolder);
-    public abstract Function<View, BaseViewHolder> getViewHolderFactory();
+    public Function<View, BaseViewHolder> getViewHolderFactory() {
+        return new ReflectionBasedViewHolderFactory(getViewHolderClass());
+    }
+
+    @SuppressWarnings("unchecked")
+    public Class<? extends TViewHolder> getViewHolderClass() {
+        return (Class<? extends TViewHolder>) ((ParameterizedType) getClass()
+                .getGenericSuperclass()).getActualTypeArguments()[0];
+    }
+
+    public static class ReflectionBasedViewHolderFactory implements Function<View, BaseViewHolder> {
+        private Class<? extends BaseViewHolder> viewHolderClass;
+
+        public ReflectionBasedViewHolderFactory(Class<? extends BaseViewHolder> viewHolderClass) {
+            this.viewHolderClass = viewHolderClass;
+        }
+
+        @Override
+        public BaseViewHolder apply(View view) {
+            try {
+                Constructor<? extends BaseViewHolder> viewHolderConstructor = viewHolderClass.getConstructor(View.class);
+                return viewHolderConstructor.newInstance(view);
+            } catch (Exception e) {
+                return null;
+            }
+        }
+    }
 }
