+++ viewcelladapter/src/main/java/ca/antonious/viewcelladapter/viewcells/GenericViewCell.java
@@ -4,7 +4,7 @@ package ca.antonious.viewcelladapter.viewcells;
  * Created by George on 2016-11-17.
  */
 
-public abstract class GenericViewCell<TViewHolder extends BaseViewHolder, TModel> extends ViewCell<TViewHolder> {
+public abstract class GenericViewCell<TViewHolder extends BaseViewHolder, TModel> extends AbstractViewCell<TViewHolder> {
     private TModel model;
 
     public GenericViewCell(TModel model) {
